pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdebench"
version = "0.1.0"
description = "Minimal PDE benchmark with FEniCSx"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
    "scipy>=1.7",
    "jsonschema>=4.0",
    "psutil>=5.8",
    "sympy>=1.10",
    "openai>=2.13",
    "boto3>=1.42",
    "google-genai>=1.59",
]

[project.optional-dependencies]
fenicsx = [
    # 安装命令: conda install -c conda-forge fenics-dolfinx=0.10.0
    "fenics-dolfinx>=0.10.0",
    "fenics-ufl>=2025.2.0",
    "mpi4py>=4.1.1",
    "petsc4py>=3.24.2",
    # 复杂几何网格生成（L形、环形、多孔等非矩形域）
    "pygmsh>=7.1.17",
    "meshio>=5.3.5",
]
agents = [
    "mini-swe-agent",
    "openhands",
]

dev = [
    "pytest>=7.0",
    "pytest-xdist>=2.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pdebench*"]
exclude = ["tests*", "scripts*", "cases*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

